meteostat/requirements.txt
meteostat==1.5.10
google-cloud-bigquery==2.28.1